# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AhBlocked": {
        "description": "Items never sold on the AH: only vendored or disenchanted.",
        "properties": {
          "details": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ItemInfo"
            },
            "title": "Details",
            "type": "object"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AhBlockedItem"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "details"
        ],
        "title": "AhBlocked",
        "type": "object"
      },
...
      "EvaluateResponse": {
        "properties": {
          "items": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ItemInfo"
            },
            "title": "Items",
            "type": "object"
          },
          "result": {
            "$ref": "#/components/schemas/RankResult"
          }
        },
        "required": [
          "result",
          "items"
        ],
        "title": "EvaluateResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SelectionModel": {
        "description": "A realm and faction: whose recipes count and which auction house prices them.",
        "properties": {
          "faction": {
            "title": "Faction",
            "type": "string"
          },
          "realm": {
            "title": "Realm",
            "type": "string"
          }
        },
        "required": [
          "realm",
          "faction"
        ],
        "title": "SelectionModel",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "altarmy-profit",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/ah-blocked/{item_id}": {
      "delete": {
        "description": "Allow selling `item_id` on the AH again.",
        "operationId": "unblock_ah_api_ah_blocked__item_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AhBlocked"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unblock Ah"
      },
      "put": {
        "description": "Never sell `item_id` on the AH: /api/rank and /api/evaluate only vendor or disenchant it.",
        "operationId": "block_ah_api_ah_blocked__item_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AhBlocked"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Block Ah"
      }
    },
...
    "/api/evaluate": {
      "post": {
        "description": "One recipe as /api/rank would give it, with the user's `choices` of sources and exit applied.",
        "operationId": "evaluate_api_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluate"
      }
    },
    "/api/game-data/update": {
      "post": {
        "operationId": "update_game_data_api_game_data_update_post",
        "parameters": [
          {
            "description": "skip the rebuild if the newest build is loaded",
            "in": "query",
            "name": "only_if_new",
            "required": false,
            "schema": {
              "default": false,
              "description": "skip the rebuild if the newest build is loaded",
              "title": "Only If New",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Game Data"
      }
    },
    "/api/rank": {
      "get": {
        "description": "What the selected realm/faction's characters can craft, most profitable first. Bounds are\ninclusive; an omitted bound is unbounded (so losses are included unless `min_profit` is set).",
        "operationId": "get_rank_api_rank_get",
        "parameters": [
          {
            "description": "rank every recipe of the characters' professions, not just learned ones",
            "in": "query",
            "name": "include_unlearned",
            "required": false,
            "schema": {
              "default": false,
              "description": "rank every recipe of the characters' professions, not just learned ones",
              "title": "Include Unlearned",
              "type": "boolean"
            }
          },
          {
            "description": "ways the crafts may be sold",
            "in": "query",
            "name": "exits",
            "required": false,
            "schema": {
              "default": [
                "vendor",
                "ah",
                "disenchant"
              ],
              "description": "ways the crafts may be sold",
              "items": {
                "enum": [
                  "vendor",
                  "ah",
                  "disenchant"
                ],
                "type": "string"
              },
              "title": "Exits",
              "type": "array"
            }
          },
          {
            "description": "copper",
            "in": "query",
            "name": "min_cost",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "copper",
              "title": "Min Cost"
            }
          },
          {
            "description": "copper",
            "in": "query",
            "name": "max_cost",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "copper",
              "title": "Max Cost"
            }
          },
          {
            "description": "copper",
            "in": "query",
            "name": "min_profit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "copper",
              "title": "Min Profit"
            }
          },
          {
            "description": "copper",
            "in": "query",
            "name": "max_profit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "copper",
              "title": "Max Profit"
            }
          },
          {
            "description": "profit / cost (0.5 = 50%)",
            "in": "query",
            "name": "min_roi",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "profit / cost (0.5 = 50%)",
              "title": "Min Roi"
            }
          },
          {
            "description": "profit / cost (0.5 = 50%)",
            "in": "query",
            "name": "max_roi",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "profit / cost (0.5 = 50%)",
              "title": "Max Roi"
            }
          },
          {
            "in": "query",
            "name": "top",
            "required": false,
            "schema": {
              "default": 50,
              "minimum": 1,
              "title": "Top",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Rank"
      }
    },
...
    "/api/selection": {
      "put": {
        "description": "Switch realm/faction; that realm's Auctionator prices replace the previous ones.",
        "operationId": "put_selection_api_selection_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectionModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Selection"
      }
    },
...
  }
}